+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 ### ✨ Improvements
 - Utils: Update Stanza's Serbian (Latin script) sentence tokenizer, part-of-speech tagger, and dependency parser
 - Utils: Update Wordless's sentence splitter and sentence segment tokenizer
+- Work Area: Parallel Concordancer - Parallel Unit No. - Empty parallel units are not counted
 
 ### 🔧 Bugfixes
 - Work Area: Fix Profiler - Syntactic Complexity